Avoid division by zero error when all values are zeroes
If all values to put on the chart are zero, then the size of the largest
"circle axis" is maxValue. But if the user did not supply a value via
configuration option, maxValue remains 0, which causes division by zero.

Set maxValue to 1 in such case. It has no affect on the chart as this
happens only when all values are zeroes.